add test and finish plugin
