+++ contracts/libraries/Helper.sol
@@ -49,4 +49,26 @@ library Helper {
     (bool success, ) = to.call{value: value}(new bytes(0));
     require(success, 'Helper::safeTransferETH: ETH transfer failed');
   }
+
+  function verifyCallResult(
+      bool success,
+      bytes memory returndata,
+      string memory errorMessage
+  ) internal pure returns (bytes memory) {
+    if (success) {
+      return returndata;
+    } else {
+      // Look for revert reason and bubble it up if present
+      if (returndata.length > 0) {
+        // The easiest way to bubble the revert reason is using memory via assembly
+
+        assembly {
+          let returndata_size := mload(returndata)
+          revert(add(32, returndata), returndata_size)
+        }
+      } else {
+        revert(errorMessage);
+      }
+    }
+  }
 }
